move pgtype_binary to pgwire as binary encoder
