--- src/utils/pgtype-binary/src/error.rs
@@ -1,41 +1,0 @@
-// Copyright 2022 Singularity Data
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-use std::fmt::Display;
-
-use serde::{de, ser};
-use thiserror::Error;
-
-/// The result of a serialization or deserialization operation.
-pub type Result<T> = std::result::Result<T, Error>;
-
-/// An error that can be produced during (de)serializing.
-#[allow(missing_docs)]
-#[derive(Error, Clone, Debug, PartialEq, Eq)]
-pub enum Error {
-    #[error("{0}")]
-    Message(String),
-}
-
-impl ser::Error for Error {
-    fn custom<T: Display>(msg: T) -> Self {
-        Error::Message(msg.to_string())
-    }
-}
-
-impl de::Error for Error {
-    fn custom<T: Display>(msg: T) -> Self {
-        Error::Message(msg.to_string())
-    }
-}
