--- src/utils/pgtype-binary/src/lib.rs
@@ -1,25 +1,0 @@
-// Copyright 2022 Singularity Data
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-//! Pgtype-binary provide a serializer to serialize the data to postgresql type binary format.
-//! It used to send to client when execute in binary format mode.
-
-#![deny(missing_docs)]
-#![feature(lint_reasons)]
-
-mod error;
-mod ser;
-
-pub use error::{Error, Result};
-pub use ser::Serializer;
